+++ html/index.html
@@ -5,8 +5,8 @@
     <title>mmt payment metadata app</title>
     
 
-    <link rel="stylesheet" href="./css/bootstrap.min.css">
-    <link rel="stylesheet" href="./css/jquery-ui.min.css">   
+    <link rel="stylesheet" href="../css/bootstrap.min.css">
+    <link rel="stylesheet" href="../css/jquery-ui.min.css">   
     <style>
       /* TODO: put this somewhere else */
       .paymentsUnfilled {
@@ -25,12 +25,12 @@
       delete module.exports
     </script>
      
-    <script src="./scripts/jquery-3.3.1.min.js"></script>
-    <script src="./scripts/jquery-ui.min.js"></script>
+    <script src="../scripts/jquery-3.3.1.min.js"></script>
+    <script src="../scripts/jquery-ui.min.js"></script>
      
     <!-- normal script imports etc  -->
      
-    <script src="./mmt-metadata.js"></script>
+    <script src="../src/mmt-metadata.js"></script>
      
     <script>
       // TODO: put this somewhere else
