+++ src/bitcoin-utils.js
