+++ src/electrum-client.js
