+++ src/main.js
@@ -15,7 +15,7 @@ function createWindow () {
   mainWindow = new BrowserWindow({width: 800, height: 600})
 
   // and load the index.html of the app.
-  mainWindow.loadURL(`file://${__dirname}/index.html`)
+  mainWindow.loadURL(`file://${__dirname}/../html/index.html`)
 
   // Open the DevTools.
   mainWindow.webContents.openDevTools()
