--- testvalidate
@@ -1,13 +1,0 @@
-
-var bitcoin = require('bitcoinjs-lib')
-var address = "mu1yJFE97UsuszLtDbqF8ci6jH3ZEFcA5uN"      
-
-function validAddress(address) {
-try {
-  return bitcoin.address.toOutputScript(address, bitcoin.networks.testnet)
-} catch (e) { return false }
-}
-
-if (valid(address)) console.log('yes') 
-  else console.log('no')
-
